Move the function for metrics to a separate script
